fix(wiki/graph): dynamic-import sigma view to avoid SSR WebGL crash
Next.js prerender was failing with `WebGL2RenderingContext is not defined`
because sigma initializes its renderer at module load. Wrap the sigma
component in `next/dynamic({ ssr: false })` so it only loads in the browser.
This is what unblocked the web container — it was crash-looping on boot.
